requirements.txt
colorama==0.4.6
PrettyPrintTree==1.36.3
